NLDQS-MCP/app/requirements.txt
fastapi
uvicorn
requests
psycopg2-binary
python-multipart
pandas
openpyxl
